Radio group editorial updates
Changed description, which described a radio button rather than a radio group. It was:
"An option in single-select list"

Moved the following note from the beginning to the end of the keyboard section, to match the editorial style of all other patterns:
<p class="note">The Tab behavior below differs from some native browser behaviors where Shift+Tab sometimes moves to the last button in the group, if none are selected.</p>\
And revised the note to read:
<p class="note">The behavior of the <kbd>Tab</kbd> key described above differs slightly from the behavior provided by some browsers for native HTML radio groups.
In some browsers, if none of the radio buttons are selected, <kbd>Shift+Tab</kbd> places focus on the last radio button instead of the first radio button. </p>

Updated the keyboard section language and format to be consistent with the APG editorial style guidelines.
Changed the language in the keyboard section to use the word "checked" instead of "selected" to avoid implying that aria-selected is used instead of aria-checked to indicate the checked state.
Reworded the space key bullet to fix issue #96.

Updated the roles, states, and properties section to be consistent with APG editorial style guidelines.