Simplify keyboard section of example coding templates
Modified files:
examples/coding-template/MultipleImplementationExample-Template.html
examples/coding-template/SingleImplementationExample-Template.html

Replaced the table with a list in the keyboard section.
Changed the comment to say that it is recommended to copy the list of keys from the design pattern in aria-practices.html
and then edit the list to describe only what is relevant to the example.