Use ordered lists for example notes
In examples/coding-template/MultipleImplementationExample-Template.html,
each example has a notes section that had an unordered list of notes.
changed to ordered list since the list items are not labeled, e.g., like key lists are labeled by the key.
This makes it possible for people to refer to a specific note, e.g., example 2, note 3.