Use ordered list for accessibility features in single example template
In examples/coding-template/SingleImplementationExample-Template.html,
there is a section with a list of accessibility features.
Changed the list from unordered to ordered.
This makes it easier for people to refer to a specific feature since the items in the list  will not often be labeled
unlike the lists of keys where every list item starts with the key name.