Remove "Guide" when referring to the title of the document from within the document.
Per issue #153, even though the document is titled WAI-ARIA Authoring Practices 1.1", there are references to the document itself as
"APG" or Authoring Practices Guide".

Rephrased the intro so the document is itself called a guide without making the word "Guide" part of the name and then adjust references to Practices Guide and APG throughout the document.