Clarify left and right in menuKeyboardOption2
In branch menuKeyboardOption2,
in aria-practices.html,
in the menu pattern keyboard section revisions for issue #120:

Edited the 3rd bullet under the descriptions of both right arrow and left arrow.
Made the language more consistent so it is easier to compare the behavior of left and right.
Described focus movement first and menu opening behavior second.