Further clarify right and left arrow keys in menuKeyboardOption2
In branch menuKeyboardOption2,
in aria-practices.html,
in the menu pattern keyboard section revisions for issue #120:

Edited the 3rd bullet under the descriptions of both right arrow and left arrow.
In the 3rd bullet, 3 actions are performed.
Converted from a sentence describing the actions to an ordered list of 3.
In action 3, made it clear that we recommend leaving focus in the menubar but give the option of moving focus into the submenu like is commonly done on Windows.