Clarify description of screen reader listbox presentation
To address the Nov 28, 2016 comment by @jnurthen on issue #122,
modified aria-practices.html:
In the listbox pattern description, changed:
When screen readers present a listbox, they render the name, state, and position of each option in the list.
To:
When screen readers present a listbox, they may render the name, state, and position of each option in the list.