Minor editorial revisions of toolbar design pattern
In toolbar design pattern section of aria-practices.html, made the following changes.
Replaced "When designing and implementing toolbars, it is strongly recommended that authors:"
with "To optimize the benefit of toolbar widgets:"
to minimize the "it is recommended" phrasing.
In the keyboard section, fixed format and phrasing to be consistent with APG editorial guidelines.
In keyboard section, added the following to the disabled controls note:
"For additional guidance, see <a href="#kbd_disabled_controls"></a>."
Added aria-labelledby to and fixed phrasing in states and properties section.