Treeview Example Warning About Computed Position and Level Properties
Per issue #109, In the treeview examples that rely on browser calculation of position and level information, added guidance to declare position and level information if testing reveals gaps in support for automatic computation.

In "File Directory Treeview Example Using Computed Properties" in file "examples/treeview/treeview-1/treeview-1a.html",
added two sentences to this paragraph:

"This example relies on the browser to compute values for aria-setsize, aria-posinset, and aria-level.
The ARIA 1.0 specification for these properties states that browsers can, but are not required to, compute their values.
So, some browser and assistive technology combinations may not compute or report correct position and level information if it is not explicitly declared.
If testing reveals gaps in support for these properties, override automatic computation by explicitly declaring their values as demonstrated in the example of a File Directory Treeview using declared properties."

In the Navigation Treeview Example Using Computed Properties in file "examples/treeview/treeview-2/treeview-2a.html",
added two sentences to this paragraph:

"This example relies on the browser to compute values for aria-setsize, aria-posinset, and aria-level.
The ARIA 1.0 specification for these properties states that browsers can, but are not required to, compute their values.
So, some browser and assistive technology combinations may not compute or report correct position and level information if it is not explicitly declared.
If testing reveals gaps in support for these properties, override automatic computation by explicitly declaring their values as demonstrated in the example of a Navigation Treeview using declared properties."