Make Breadcrumb Example Link Separators With CSS Boarders
Per discussion in issue #187, the visual separators between breadcrumb links are only for visual presentation and create unnecessary verbosity if they are announced by screen readers. This change displays them with CSS boarders so they are not announced.

This commit also includes revisions to the accessibility features section of the breadcrumb example page to describe this feature.