Layout Grid Example: Show Keyboard NUX only if Grid Focused With Tab Key
Introduce an enhancement for the layout grid example review in issue #227.

The first layout grid provides a NUX to help new keyboard users understand that arrow keys can be used to navigate. This change will cause the NUX to be displayed only if the grid receives focus via the tab key.