Modal Dialog Pattern: Add aria-hidden guidance
For issue #325, modified the modal dialog design pattern section of aria-practices.html to add:

Optionally, if content outside a dialog is completely inert and visually obscured to an extent that is intentionally unreadable,
each element containing a portion of the inert layer has aria-hidden set to true.
In this circumstance, the dialog container element cannot be a descendant of an element that has aria-hidden set to true.
However, if content outside a modal dialog is visually discernable, aria-hidden is not present.