Dialog Pattern: Clarify Description of Focus Movement
I was reminded that some focusable elements are not in the tab sequence so the description of the keyboard behavior is not clear because it says that tab moves focus to next focusable element.

So, for issue #325, made the following changes.

Added the following paragraph to the beginning of the keyboard subsection:

> In the following description, the term “tabbable element” refers to any element with a tabindex value of zero or greater.
> Note that values greater than 0 are strongly discouraged.

Added the following to the keyboard interaction description as the first item in the list:

> When a dialog opens, focus moves to an element inside the dialog. See notes below regarding initial focus placement.

In the subsequent bullet items in the keyboard description, replaced the word "focusable" with "tabbable".