Remove Incomplete Presentation Role Examples Page
Work on issue #182 is not sufficiently complete for release 1.
However, a rough draft of content had been merged into master.
Removing the rough draft file and the link to it.

The rough draft file and the linking paragraph will be added back in a new feature branch called issue182-presentation-examples.
When work on the presentation examples is complete, we will merge that feature branch back to master.

Changes:

1. Modified aria-practices.html to remove link to presentation examples page.
2. deleted examples/presentation/PresentationRoleExamples.html