Layout Grid Examples: Fix problems described in issue 758
modified examples/grid/LayoutGrids.html
to fix the following problems described in issue #758.

Problem 1: tabindex="-1" is not only applied to span elements. It is applied to div as well.

Resolution 1:
Updated documentation of tabindex.
Made it consistent with descriptions of roving tabindex in other examples.
Added documentation of tabindex=0 as well.

Problem 2: The definition of the word "cell" is not consistent between the attribute table and the keyboard table.

Resolution 2: Added a note above the keyboard table describing how cell is used.