Generated by TRAVIS-CI 499238870bacf82a6943079e24692db799fa8149
Spin Button Pattern: Add aria-invalid guidance for values outside allowed range (pull #825)

In response to issue #704, modified the roles, states, and properties section of the spin button pattern in aria-practices.html.

Added a bullet to the list of roles, states, and properties stating that aria-invalid is set true if it is possible for the textbox to contain a value that is not allowed.