Generated by TRAVIS-CI bf220754faecffb96ed0dd528237220016aab1ac
Naming and describing section: Remove bibliography reference (pull#1059)

PR #994 included changes to common/biblio.js that should not have been merged
because that file is maintained outside the aria-practices repository.
The change was addition of a reference to an external site.
This commit removes the common/biblio.js changes and the reference to them.