Generated by TRAVIS-CI ecdb87a55dceed94ce0460bafc80d99c96757f3a
Date Picker Dialog: fixes bug when clicking on a disabled day (pull #1095)

In the calendar grid in the date picker dialog, days that are in a previous or next month can appear in the grid.
These are disabled. Clicking them generated a JS error.
This fix causes the month displayed in the grid to change to the month that includes the day clicked.