Generated by TRAVIS-CI e653a7afd9501a4649061f609dbe28e7a68f78a7
Date Picker Dialog Example: hide dates from prior or next month (pull #1122)

Dates at the beginning and end of the grid that are for days in the prior or next month were displayed and disabled.
To resolve issue #1121, this commit removes those dates from the grid.
It also removes aria-selected from the button representing the date currently displayed in the edit field because aria-selected is not valid on button.