Generated by TRAVIS-CI 9a2bbb77e2fc5ce5272aff4a4a706cea98ed179f
Update master with all 1.2-specific changes from the apg-1.2 branch (pull#1254)

Makes the master branch the editor's draft for APG 1.2; it was previously editor's draft for APG 1.1.
Merges all commits from the apg-1.2 branch that were not in master into master except that it leaves out the change history portion of the appendix.