Generated by TRAVIS-CI eec5af60833c7d488a3874af3246641f68609014
Slider patterns: Provide warning about inability to operate with touch-based assistive technologies due to lack of needed APIs (pull #1186)

Addresses issue #1150 for the slider patterns by adding a warning.
The warning explains that widgets using the pattern are not fully operable with touch-based assistive technologies
because the needed APIs do not yet exist.

Co-authored-by: Matt King <a11yThinker@Gmail.com>
Co-authored-by: Carolyn MacLeod <Carolyn_MacLeod@ca.ibm.com>