Generated by TRAVIS-CI 5161cf11a4c3e48e601cbc51b3d71e8b323e7464
Tree Examples: Correct result of selecting parent node (pull #1306)

Resolves #760.

The file selector tree view examples display the selected tree item in a neighboring read-only text input.
This PR changes what is displayed in the text input when a parent tree item is selected.
Previously, it displayed the parent node treeitem element, showing HTML markup for a span contained by the parent node.
Now, when a parent node is selected, it displays innerText from the span that names the parent item.
For end nodes, it continues to use inner text of the tree item node for the selected item display.

Co-authored-by: Matt King <a11yThinker@Gmail.com>