Generated by TRAVIS-CI 3909695fa0f136c4b55989aef7106e1d4c044923
utils.js: Make isFocusable return false when tabindex is positive and element is disabled (pull#1419)

This change potentially effects any example that uses the isFocusable method in the shared utility scripts (examples/js/utils.js).
Fixes a bug in isFocusable where it returned true for elements with positive tabindex even when the element is disabled.