fix listbox group label issue
This ensures that the label of a group in a listbox is not counted / identified by AT as also being a member of the group. Putting role="presentation" on the li just means the li will be ignored by AT, but not its children. The text node instead ends up as a member of the group.