fix(#2238): adds catch for when control number is 0 (#2242)
The example wraps. So in certain scenarios the control number will be 0. For example, 25th of February “minus” 31 days results in the code example to land on day 28. Due to the module on line 358 that would make the control number 0. This catches that and sets it to dayMax (same logic as in the code example itself).