Feed Example: Convert images to inline SVG (#2729)
Changes the images used for star ratings of restaurants in the feed example to inline SVG.
Because the build process that converts to Jekyll for the production site separates images into a separate path, the links to the dynamically loaded image assets were being broken by the build process. With this change, the JavaScript is no longer dependent on paths to image files.