Infrastructure: Change localhost references to accommodate Node V17 (#2749)
Regression tests failed to run with Node V17, so we temporarily pinned the version to 16. This is an alternative fix that enables running with node V17+.

This commit unpins the node version and replaces references to localhost with 127.0.0.1.

Node v17+ deprioritised IPv4 addresses in favour of IPv6 with a change to dns.setDefaultResultOrder.
Node made this change without a happy-eyeballs implementation, which meant the loopback address used when running the regression tests resolved to ::1 rather than the IPv4 loopback (127.0.0.1) -- the address geckodriver expected.
So this PR now explicitly uses 127.0.0.1.
Alternatively, we could have added --dns-result-order=ipv4first to the running script, or added dns.setDefaultResultOrder('ipv4first');, but would eventually remove it as the happy-eyeballs implementation is now included in v20.