Correct text for Control + Home/End in treegrid pattern
The text in the treegrid pattern differed to the text in the treegrid
example. The content of the treegrid pattern was nonsensical as a column
only contains one cell so moving focus to the first cell in a column is
meaningless.

The behaviour and text of the example are consistent and make logical
sense.

Inaccurate text:
If focus is on a cell, moves focus to the first cell in the column.
If focus is on a cell, moves focus to the last cell in the column.

Corrected text:
If focus is on a cell, moves focus to the cell in the first row in the same column as the cell that had focus.
If focus is on a cell, moves focus to the cell in the last row in the same column as the cell that had focus.