Treegrid Pattern keyboard section: Correct text for Control + Home/End (pull #2779)
For Control+Home and Control+End, The text in the treegrid pattern differed from the text in the treegrid
example. The text in the treegrid pattern didn't make sense because a column only contains one cell so moving focus to the first cell in a column is meaningless. The behaviour and text on the example page are consistent and make logical
sense. This commit makes the pattern and example pages consistent with each other.

Inaccurate text:
If focus is on a cell, moves focus to the first cell in the column.
If focus is on a cell, moves focus to the last cell in the column.

Corrected text:
If focus is on a cell, moves focus to the cell in the first row in the same column as the cell that had focus.
If focus is on a cell, moves focus to the cell in the last row in the same column as the cell that had focus.