Infrastructure: Source heading for read this first content from aria-practices repo (pull #2867)
Moves the h2 element that is part of the "Read This First" content from the wai-aria-practices repo into the aria-practices repo. This is part of an effort to ensure all content is sourced and controlled from the aria-practices repo and the wai-aria-practices repo provides only transformations necessary for site build.