Infrastructure: Fix inconsistencies in lists of CSS and JS files on 3 example pages (Related to wai-aria-practices issue 255) (pull #2883)
Each example page has a list of CSS and JS files used by the sourceCode function in shared/js/examples.js. Three examples had errors in this list that are corrected by this commit. This fix supports resolution of w3c/wai-aria-practices#255, which is related to incorrect last revision dates being shown in the footer of example pages.