Infrastructure: Transition revision date generation for Coverage and Quality Report to build repository workflow (pull #2976)
Makes the last updated date on the coverage and quality report the date the revised report it is merged to main in the content repository. The date is generated and added to the report during the build process that is executed in the build repo (wai-aria-practices). This is consistent with how the last update dates for examples are determined.

Note that the report is re-generated by the workflow in the content repo for all pull requests.