All Example Pages: Update links to related Issues (#3123)
Fixes issue #3113.

The links to "related issues" on each example page point to a GitHub project. There is a seprate GitHub project for each pattern or group of similar patterns.

The sunset of classic GitHub projects resulted in redirects from classic project URLs to the new migrated project URLs. This PR changes the links to point directly to the new project pages.

---------

Co-authored-by: Matt King <a11yThinker@gmail.com>