Infrastructure: Address failing macOS regression tests (#3115)
Adds two small utility methods: 'isMacOS()' and 'translatePlatformKeys()'. The latter utility takes a single Key or an array of Keys and returns the keys with any platform-specific Keys translated. For example, Keys.Control becomes Keys.Meta on macOS.

These utilities are then used in applicable tests to ensure that the OS-specific modifier key is used.