Developing a Keyboard Interface Practice: Add note about pointer activation and roving tabindex/aria-activedescendant (pull #3311)
Fixes #1710:
Adds a new note advising authors to change the roving tabindex  or set the correct aria-activedescendant reference as a result of pointer interactions (clicks, taps, ...) as otherwise there will be a mismatch for users that navigate with both keyboard and pointer.

Also adds <code> ... </code> around relevant words like tabindex and aria-activedescendant.

---------
Co-authored-by: Patrick H. Lauke <redux@splintered.co.uk>
Co-authored-by: Ariella Gilmore <ariellalgilmore@gmail.com>